Menu works-ish
There's a line in imagetargets.cpp that you can use to test it early.

Problems: can't update the level number for some reason.
Can't go back to the game...

+++ jni/UnitMethods.h
@@ -68,6 +68,6 @@ void animateEnemy(QCAR::Matrix44F& enemyMatrix, int enemyNumber, int x_offset, i
 void updateMissileDefaultPos(int missileNumber, float lx, float ly);
 void removeEnemy (int enemyNumber);
 void gameOver ();
-void moveEnemy (int &x, int &y, int &direction, int speed, float timeDiff);
+void moveEnemy (float &x, float &y, float &direction, float speed, float timeDiff);
 
 #endif // _UNITDB_H_
